feat: 重构项目架构并添加新功能
- 重构项目结构，采用分层架构设计
- 添加zustand状态管理
- 实现缓存工具和frontmatter解析工具
- 新增音乐页面和主题切换功能
- 优化博客页面和关于页面UI
- 添加布局组件和核心App组件
- 更新项目架构文档
