+++ package.json
@@ -22,7 +22,8 @@
     "react-router-dom": "^7.10.1",
     "tailwind-merge": "^3.4.0",
     "three": "0.160.0",
-    "tone": "^14.7.77"
+    "tone": "^14.7.77",
+    "zustand": "^5.0.9"
   },
   "devDependencies": {
     "@types/node": "^22.14.0",
